% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmcDementiaPredictionBase.R
\docType{package}
\name{EmcDementiaPredictionBase}
\alias{EmcDementiaPredictionBase}
\title{EmcDementiaPredictionBase}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
